# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domains.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      updated_at:
        type: string
    type: object
  dto.BoolResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        example: true
        type: boolean
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
      txn_id:
        example: txn_123456
        type: string
    type: object
...
  dto.EmptyResponse:
    properties:
      code:
        example: 0
        type: integer
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
      txn_id:
        example: txn_123456
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.MarketplaceCredentialRequest:
    properties:
      app_id:
        type: string
      app_key:
        type: string
      app_secret:
        type: string
      platform:
        enum:
        - shopee
        - lazada
        type: string
      sign_method:
        type: string
      user_token:
        type: string
    required:
    - platform
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.StringResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        example: data_value
        type: string
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
      txn_id:
        example: txn_123456
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/domains.User'
      message:
        example: User retrieved successfully
        type: string
      success:
        example: true
        type: boolean
      txn_id:
        example: txn_123456
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing marketplace affiliate links (Lazada, Shopee)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Market Place Affiliate API
  version: "1.0"
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return session token
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StringResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
      summary: Login user
      tags:
      - user
...
  /user/market-credential:
    post:
      consumes:
      - application/json
      description: Save or update marketplace API credentials (Lazada/Shopee)
      parameters:
      - description: Marketplace credential
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MarketplaceCredentialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Save marketplace credentials
      tags:
      - user
  /user/market-credential/{platform}:
    delete:
      description: Delete stored marketplace credentials for a platform
      parameters:
      - description: Platform name (lazada/shopee)
        in: path
        name: platform
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete marketplace credentials
      tags:
      - user
    get:
      description: Check if marketplace credentials exist for a platform
      parameters:
      - description: Platform name (lazada/shopee)
        in: path
        name: platform
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BoolResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Check marketplace credentials
      tags:
      - user
  /user/me:
    get:
      description: Get authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - user
  /user/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Register request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StringResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
      summary: Register a new user
      tags:
      - user
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"